Remove commands without code blocks (#70)
* Remove commands without code blocks

* Fix tests for commands with no scripts

* command can exist without executor

This is a design decision. It is possible that the user might have
missed giving the executor for a command. In such a case it is better
to retain the command and provide a helpful error message indicating
the problem. Telling the user that the executor is missing is better
than saying the command does not exist (because it has been pruned)

* Fix typo

* Use is_empty to check for source