Fix issue where command before level 1 heading was ignored
The command before the level 1 heading wasn't being pushed into the array of commands (line 26 in the parser).

However, after the parser is done, the last command is correctly added to the list of commands. So I decided that if we see another level one heading within the document, the parser should stop (break) there.