Use triple backticks for markdown snippets in `README.md`
This is an æsthetic/subjective nit, so feel free to dismiss it. Nonetheless, I do find that using `~~~` for code blocks in Markdown to be an extremely rare sight (in fact, TIL!), especially in the context of markdown-masquerading tools such as Discord, which do _not_ interpret `~~~` as introducing a code block. What this, in turn, entails, is that people coming across your tool will run into a bit more strangeness / impedance, and as we all know it, the budget for it is limited, and for some by quite a bit.
Doing the "typical" thing, albeit boring, ought to help in this regard :)